scripts/requirements.txt
nba_api>=1.4.1
psycopg2-binary>=2.9.9
pandas>=2.0.0
